backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
pydantic
python-dotenv
requests
google-api-python-client
langchain-core
huggingface_hub